chore: rename, make exit code constant
