+++ server/src/defines.js
@@ -10,3 +10,4 @@ export const FEE = 0.5; // Transfer (Type 0) Tx fee;
 export const MIN_PAYOUT = 0.51; // check minpayout in config to be not less; than 0.51 ADM;
 export const EPOCH = Date.UTC(2017, 8, 2, 17, 0, 0, 0); // ADAMANT's epoch time
 export const FORMAT_PAYOUT ='yyyy-MM-dd';
+export const EXIT_CODE_ERROR = 1;
